test_requirements.txt
coverage
fabric
flake8
ipdb
mixer
mock
tox
django-compat
six
pillow
